+++ src/azure-firewall/azext_firewall/profiles.py
@@ -5,4 +5,4 @@
 
 from azure.cli.core.profiles import CustomResourceType
 
-CUSTOM_FIREWALL = CustomResourceType('azext_firewall.vendored_sdks.v2020_05_01', 'NetworkManagementClient')
+CUSTOM_FIREWALL = CustomResourceType('azext_firewall.vendored_sdks.v2020_07_01', 'NetworkManagementClient')