+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,17 @@
 
 Release History
 ===============
+0.4.66
++++++
+* Add support for GitOps as an AKS addon
+* Update standard load balancer (SLB) max idle timeout from 120 to 100 minutes
+
+0.4.65
++++++
+* Honor addon names defined in Azure CLI
+* Add LicenseType support for Windows
+* Remove patterns for adminUsername and adminPassword in WindowsProfile
+
 0.4.64
 +++++
 * Add support for Open Service Mesh as an AKS addon